rename xfel sases to aircoils only. Add SASE3
