+++ environments/xfel_sase1_aircoils/configs.yaml
@@ -1,0 +1,13 @@
+---
+name: xfel_sase1_aircoils
+version: "0.2"
+dependencies:
+  - numpy
+  - badger-opt
+interface:
+  - doocs
+  - doocs_mock
+  - doocs_real2sim
+params:
+  - waiting_time: 1
+  - interface: doocs
