+++ environments/xfel_sase2_aircoils/configs.yaml
