+++ environments/xfel_sase3_aircoils/README.md
@@ -1,0 +1,5 @@
+# XFEL_SASE2 Environment for Badger
+
+## Prerequisites
+
+## Usage
